pyproject.toml
[tool.poetry]
name = "pear-back"
version = "0.1.0"
description = ""
authors = ["Mmarcos01 <marie.marcos25@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "^3.2.9"
Pillow = "^8.4.0"
django-cors-headers = "^3.10.0"
djangorestframework-simplejwt = "^5.0.0"
djangorestframework = "^3.12.4"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
